finalized scripts for artifacts

+++ scripts/runAllPairs_monkey.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 
-#(test $# -lt 1) && (echo "too few arguments") && exit 0
+(test $# -lt 3) && (echo "too few arguments") && exit 0
 
-tmv=${1:-"3600"}
-OUTDIR=/home/hcai/testbed/monkeyLogs_explicit
-#OUTDIR=/home/hcai/testbed/monkeyLogs_implicit
+indir=$1
+linkage=$2
+tmv=${3:-"3600"}
+srcdir=/home/hcai/testbed/cg.instrumented/$indir/pairs/${linkage}
+OUTDIR=/home/hcai/testbed/pairTrace_$linkage
 mkdir -p $OUTDIR
 
 timeout() {
@@ -23,17 +25,23 @@ timeout() {
 
 }
 
-#for ((i=1;i<=54;i++))
-for ((i=1;i<=7;i++))
+for ((i=1;i<=55;i++))
 do
-	echo "================ RUN APP PAIR $i ==========================="
-	/home/hcai/testbed/setupEmu.sh Galaxy-Nexus-23
-	/home/hcai/testbed/singlePairInstall.sh $i
-	adb logcat -v raw -s "hcai-intent-monitor" "hcai-cg-monitor" &>$OUTDIR/${i}.logcat &
-	#adb shell monkey --ignore-crashes --ignore-timeouts --ignore-security-exceptions --throttle 200 1
-	timeout $tmv "/home/hcai/testbed/runPair.sh $i >$OUTDIR/${i}.monkey"
+	if [ ! -s $srcdir/${i}/s.apk ];then continue; fi
+	if [ ! -s $srcdir/${i}/t.apk ];then continue; fi
 
-	adb kill-server
+	echo "================ RUN APP PAIR $i ==========================="
+	echo "Starting android emulator ......"
+	/home/hcai/testbed/setupEmu.sh Nexus-One-10 2>/dev/null 1>&2
+
+	/home/hcai/testbed/singlePairInstall.sh $indir $linkage $i
+	adb logcat -v raw -s "hcai-intent-monitor" "hcai-cg-monitor" &>$OUTDIR/${i}.logcat &
+
+	timeout $tmv "/home/hcai/testbed/runPair.sh $indir $linkage $i >$OUTDIR/${i}.monkey"
+
+	/home/hcai/testbed/singlePairUninstall.sh $indir $linkage $i
+
+	killall -9 adb
 	killall -9 adb
 done
